server/requirements.txt
flask
flask-cors
torch
torchvision
timm
pillow
numpy
onnx
onnxruntime
python-multipart
opencv-python
tqdm
requests
opencv-contrib-python
